FloatingPoint: Added workarounds for the remaining problems in pow.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/fix_floatingpoint_calculator@46174 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/floatingpoint/CMyFloat.java
@@ -9,6 +9,7 @@
 package org.sosy_lab.cpachecker.util.floatingpoint;
 
 import com.google.common.base.Preconditions;
+import java.math.BigInteger;
 import org.kframework.mpfr.BigFloat;
 import org.kframework.mpfr.BinaryMathContext;
 import org.sosy_lab.cpachecker.util.floatingpoint.CFloatNativeAPI.CNativeType;
@@ -170,7 +171,7 @@ public class CMyFloat extends CFloat {
 
   @Override
   public CFloat powToIntegral(int exponent) {
-    return new CMyFloat(delegate.powInt(exponent));
+    return new CMyFloat(delegate.powInt(BigInteger.valueOf(exponent)));
   }
 
   @Override
